+++ app.js
@@ -8,12 +8,12 @@ import militarySymbolsObject from './militarySymbolsObject';
 import unitSizeObject from './unitSizeObject';
 import mod1Object from './mod1Object';
 import mod2Object from './mod2Object';
-
+import taskForceObject from './taskForceObject';
 
 // * The star of the show * //
 // ex- new MilSym('.test', 'Infantry', 'friendly', 'team', 'Armored', 'Rail', 'A/2-101', '27/42ID', '+', false, false, false).placeSymbol();
 class MilSym {
-  constructor(location, symbol, affiliation = 'friendly', echelon = 'none', mod1 = 'None', mod2 = 'None', uniqueDesignation = '', higherFormation = '', reinforcedReduced = '', isFlying = false, isActivity = false, isInstallation = false) {
+  constructor(location, symbol, affiliation = 'friendly', echelon = 'none', mod1 = 'None', mod2 = 'None', uniqueDesignation = '', higherFormation = '', reinforcedReduced = '', isFlying = false, isActivity = false, isInstallation = false, isTaskForce = false) {
     this.location = document.querySelector(location);
     this.symbol = militarySymbolsObject[symbol].affiliation[affiliation];
     this.affiliation = affiliationOutlineObject[affiliation];
@@ -28,6 +28,7 @@ class MilSym {
     this.flightCapable = militarySymbolsObject[symbol].flightCapable;
     this.isActivity = isActivity;
     this.isInstallation = isInstallation;
+    this.isTaskForce = isTaskForce;
     this.data = {
       location,
       symbol,
@@ -58,7 +59,7 @@ class MilSym {
     // } else {
     //   svg.append(this.affiliationOutlineData, this.decoratorData, this.echelonData);
     // }
-    svg.append(this.affiliationOutlineData, this.decoratorData, this.echelonData, this.mod1Data, this.mod2Data, this.uniqueDesignationData, this.higherFormationData, this.reinforcedReducedData, this.activityData, this.installationData);
+    svg.append(this.affiliationOutlineData, this.decoratorData, this.echelonData, this.mod1Data, this.mod2Data, this.uniqueDesignationData, this.higherFormationData, this.reinforcedReducedData, this.activityData, this.installationData, this.taskForceData);
     this.location.append(svg);
     svg.setAttributeNS(null, 'data-symbol-name', this.data.symbol);
     svg.setAttributeNS(null, 'data-symbol-info', JSON.stringify(this.data)); // this should probably be split into separate data-attrs
@@ -444,7 +445,7 @@ class MilSym {
   get installationData() {
     if (this.isInstallation) {
       const installationGroup = document.createElementNS('http://www.w3.org/2000/svg', 'g');
-      installationGroup.classList.add('activity');
+      installationGroup.classList.add('installation');
       const installationModifier = document.createElementNS('http://www.w3.org/2000/svg', 'path');
       installationModifier.setAttribute('d', this.affiliation.installation);
       installationModifier.setAttribute('fill', 'black');
@@ -452,6 +453,20 @@ class MilSym {
       installationModifier.setAttribute('stroke-width', '4');
       installationGroup.append(installationModifier);
       return installationGroup;
+    }
+  }
+
+  get taskForceData() {
+    if (this.isTaskForce) {
+      const taskForceGroup = document.createElementNS('http://www.w3.org/2000/svg', 'g');
+      taskForceGroup.classList.add('taskforce');
+      const taskForceModifier = document.createElementNS('http://www.w3.org/2000/svg', 'path');
+      taskForceModifier.setAttribute('d', taskForceObject[this.data.echelon].affiliation[this.data.affiliation].d);
+      taskForceModifier.setAttribute('fill', 'none');
+      taskForceModifier.setAttribute('stroke', 'black');
+      taskForceModifier.setAttribute('stroke-width', '4');
+      taskForceGroup.append(taskForceModifier);
+      return taskForceGroup;
     }
   }
 }